requirements.txt
numpy
scikit-fuzzy
pillow
